% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_axis.R
\name{ActiveIdent}
\alias{ActiveIdent}
\title{Activate specific clusters for analysis}
\usage{
ActiveIdent(senna, id)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{id}{A vector of cluster identities to activate}
}
\value{
A modified `SENNA` object with activated cluster information
}
\description{
Reflects cluster label information in the analysis based on the saved annotation.
}
