% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_senna.R
\name{AddReference}
\alias{AddReference}
\title{Add cell attribute}
\usage{
AddReference(senna, var_name, reference)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{var_name}{Name of the cell attribute}

\item{reference}{A named vector of attribute values (must match cell names)}
}
\value{
A modified `SENNA` object with the new cell attribute added
}
\description{
Adds a cell attribute (e.g., annotation or metadata) to the `SENNA` object.
}
