% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_senna.R
\name{AppDat}
\alias{AppDat}
\title{Creating app data}
\usage{
AppDat(
  senna,
  reference_value = NULL,
  colorset = NULL,
  image_path = NULL,
  image_resolution = c("lowres", "hires"),
  envir = globalenv()
)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{reference_value}{Name of reference data to visualize (e.g., annotation or numeric vector)}

\item{colorset}{If `reference_value` is discrete, a named vector of colors; if continuous, a palette}

\item{image_path}{Path to the Visium image folder. If NULL, no image will be included.}

\item{image_resolution}{Image resolution to use: either `"lowres"` or `"hires"`}

\item{envir}{Environment where `app_dat` and `app_image` will be assigned}
}
\value{
A data frame named `app_dat` (in `envir`)
}
\description{
Prepares coordinate and annotation data (and optionally image data) from a `SENNA` object for interactive app use.
}
