% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_senna.R
\name{ConvertMultiSENNA}
\alias{ConvertMultiSENNA}
\title{Create mSENNA object from multiple SENNA objects}
\usage{
ConvertMultiSENNA(SENNA_list)
}
\arguments{
\item{SENNA_list}{A list of `SENNA` objects}
}
\value{
A `mSENNA` object
}
\description{
Constructs a `mSENNA` object by combining multiple `SENNA` objects.
}
