% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_senna.R
\name{CropSENNA}
\alias{CropSENNA}
\title{Crop SENNA}
\usage{
CropSENNA(senna, xlim = c(0, 1), ylim = c(0, 1), scaling = FALSE)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{xlim, ylim}{Ranges to crop. Each should be a two-length scaled vector indicating minimum and maximum values}

\item{scaling}{Logical. If TRUE, performs 0–1 scaling on each coordinate}
}
\value{
A cropped `SENNA` object
}
\description{
Crops the field of view (FOV) of a `SENNA` object.
}
