% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{crvtrjrylin}
\alias{crvtrjrylin}
\title{Generate straight-line curve trajectory using piecewise linear interpolation}
\usage{
crvtrjrylin(senna)
}
\arguments{
\item{senna}{A `SENNA` object with a straight-type curve axis}
}
\value{
A data frame of 2D coordinates sampled along the straight-line trajectory
}
\description{
Compute intermediate spatial coordinates along a curve axis when  
it is defined as a straight path. This method uses piecewise linear  
interpolation of the precomputed coefficients.
}
