% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{crvtrjryspl}
\alias{crvtrjryspl}
\title{Generate spline-based curve trajectory}
\usage{
crvtrjryspl(senna)
}
\arguments{
\item{senna}{A `SENNA` object containing spline curve axis information}
}
\value{
A data frame of 2D coordinates sampled along the spline trajectory
}
\description{
Generate a dense set of spatial coordinates along a cubic spline curve  
interpolated through the user-defined knots. If the curve is not trimmed,  
linear extrapolation is added at both ends using precomputed curve parameters.
}
