% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_axis.R
\name{ctype_linear}
\alias{ctype_linear}
\title{Full curve (piecewise linear)}
\usage{
ctype_linear(senna, knot_df, t)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{knot_df}{A data frame of knots}

\item{t}{A numeric vector specifying knot positions}
}
\value{
A `SENNA` object with linear curve axis stored
}
\description{
Computes piecewise linear curve axis (CA) coefficients from given knots.
}
