% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_axis.R
\name{ctype_spline}
\alias{ctype_spline}
\title{Full curve (spline type)}
\usage{
ctype_spline(senna, knot_df, t)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{knot_df}{A data frame of knots}

\item{t}{A numeric vector specifying knot positions}
}
\value{
A `SENNA` object with spline curve axis stored
}
\description{
Computes spline-based curve axis (CA) coefficients from given knots.
}
