% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{derpolcoef}
\alias{derpolcoef}
\title{Differentiate curve axis coefficients}
\usage{
derpolcoef(f.list)
}
\arguments{
\item{f.list}{A list containing `t`, `x.coef`, and `y.coef` from a spline curve axis}
}
\value{
A list of differentiated coefficients (`x.deriv`, `y.deriv`) and the corresponding trimmed `t` vector
}
\description{
Computes first derivatives of x and y polynomial coefficients from a spline-based curve axis (CA).
}
