% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_axis.R
\name{FullCurve,SENNA-method}
\alias{FullCurve,SENNA-method}
\title{Create a curve axis (CA) by connecting the given knots with extrapolation}
\usage{
\S4method{FullCurve}{SENNA}(senna, knot_df, type = "spline")
}
\arguments{
\item{senna}{A `SENNA` object}

\item{knot_df}{A data frame of 2D coordinates representing the curve knots}

\item{type}{A string indicating the curve type; one of `"spline"` or `"straight"`}
}
\value{
A modified `SENNA` object with a fitted full-length curve
}
\description{
Defines the method for `FullCurve()` when applied to a `SENNA` object. This function constructs a full curve axis by connecting the input knots and optionally extends the curve beyond the first and last knots. This is suitable for full-field 'progression' or 'regionation' scenarios.
}
