% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_axis.R
\name{FullCurve}
\alias{FullCurve}
\title{Create a curve axis (CA) by connecting the given knots with extrapolation}
\usage{
FullCurve(senna, ...)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{...}{Additional arguments}
}
\description{
Defines the generic function `FullCurve()`. It is suitable for both 'regionation' scenarios and 'progression' scenarios where the entire FOV is used.
}
