% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{GetCurveParam}
\alias{GetCurveParam}
\title{Compute curve parameter}
\usage{
GetCurveParam(senna, cores = 1L)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A `SENNA` object with updated `Spatial` coordinates containing curve parameter
}
\description{
Computes curve parameter values (e.g., t) for each spot based on the curve axis (CA) type in the `SENNA` object.
}
