% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{getcvrtisl}
\alias{getcvrtisl}
\title{Get islet curve parameter}
\usage{
getcvrtisl(senna, cores)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A data frame of updated spatial coordinates with curve parameter values
}
\description{
Computes curve parameter values for each spot based on a closed islet-type curve axis (CA).
}
