% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{getcvrtlin}
\alias{getcvrtlin}
\title{Get linear curve parameter}
\usage{
getcvrtlin(senna, trimmed, cores)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{trimmed}{Logical. Whether the curve is trimmed}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A data frame of updated spatial coordinates with curve parameter values
}
\description{
Computes curve parameter for each spot based on a linear curve axis (CA). Supports both trimmed and untrimmed modes.
}
