% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{getcvrtspl}
\alias{getcvrtspl}
\title{Get spline curve parameter}
\usage{
getcvrtspl(senna, trimmed, cores)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{trimmed}{Logical. Whether the spline is trimmed}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A data frame of updated spatial coordinates with curve parameter values
}
\description{
Computes curve parameter for each spot based on a spline curve axis (CA). Handles both trimmed and untrimmed curves.
}
