% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{GroupCurveCovariate,SENNA-method}
\alias{GroupCurveCovariate,SENNA-method}
\title{Grouping curve parameter}
\usage{
\S4method{GroupCurveCovariate}{SENNA}(senna, bins = NULL)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{bins}{An integer vector or number of breakpoints for binning}
}
\value{
A `SENNA` object with binned curve parameter added to `Spatial`
}
\description{
Groups spatial spots based on their curve parameter (`t`) into discrete bins for downstream analysis.
}
