% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{GroupCurveCovariate}
\alias{GroupCurveCovariate}
\title{Group curve parameter for downstream analysis}
\usage{
GroupCurveCovariate(senna, ...)
}
\arguments{
\item{senna}{A `SENNA` or `mSENNA` object}

\item{...}{Additional arguments}
}
\value{
A plot highlighting grouped regions along the curve
}
\description{
Defines the generic function `GroupCurveCovariate()`, used to group spots along the curve parameter for downstream analysis.
}
