% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{hfun}
\alias{hfun}
\title{Evaluate h(t) polynomial terms}
\usage{
hfun(xcoef, xder, ycoef, yder, cores = 1L)
}
\arguments{
\item{xcoef}{A list of x polynomial coefficients}

\item{xder}{A list of x polynomial derivative coefficients}

\item{ycoef}{A list of y polynomial coefficients}

\item{yder}{A list of y polynomial derivative coefficients}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A list of h(t) polynomial expressions per segment
}
\description{
Computes h(t) = x(t)x'(t) + y(t)y'(t) for each segment using polynomial coefficients and their derivatives.
}
