% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{isl_roots}
\alias{isl_roots}
\title{Compute curve parameter and distance for each spot (islet curve)}
\usage{
isl_roots(spatial, hfun, xcoef, xder, ycoef, yder, init, cores = 1L)
}
\arguments{
\item{spatial}{A data frame of spatial coordinates}

\item{hfun}{A list of h(t) polynomial terms for each curve segment}

\item{xcoef}{Polynomial coefficients for x(t)}

\item{xder}{Derivatives of x(t)}

\item{ycoef}{Polynomial coefficients for y(t)}

\item{yder}{Derivatives of y(t)}

\item{init}{The initial knot used to close the curve}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A data frame with additional columns: `t` (curve parameter) and `distance` (Euclidean distance to curve)
}
\description{
Computes the curve parameter (t) and Euclidean distance from each spatial spot to an islet-type closed curve axis (CA), using root solving.
}
