% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_senna.R
\name{knot_picker}
\alias{knot_picker}
\title{Run knot-picker app}
\usage{
knot_picker(path = system.file("shiny/knot_picker", package = "SENNA"))
}
\arguments{
\item{path}{App path. Defaults to the bundled app within the package}
}
\description{
Launches the interactive Shiny app for selecting knots on a spatial map.
}
