% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGs_tools.R
\name{midpprog}
\alias{midpprog}
\title{Generics for Progression Peak Detection Model}
\usage{
midpprog(senna, ...)
}
\arguments{
\item{senna}{A `SENNA` or `mSENNA` object}

\item{...}{Additional arguments passed to method implementations}
}
\description{
Define the generic function `midpprog()` for identifying genes with peak-like expression  
patterns along the progression axis.
}
