% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGs_tools.R
\name{midpregi,SENNA-method}
\alias{midpregi,SENNA-method}
\title{Regionation Peak Detection, SENNA-method}
\usage{
\S4method{midpregi}{SENNA}(
  senna,
  pk,
  subj,
  FDR_level = 0.01,
  grad_cutoff = 0.1,
  active = FALSE,
  direction = NULL,
  regress_out_cc = FALSE,
  cores = 1L
)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{pk}{A numeric value or vector indicating the peak position(s)}

\item{subj}{A character vector of genes to test}

\item{FDR_level}{FDR threshold (default = 0.01)}

\item{grad_cutoff}{Threshold for the absolute value of regression coefficients}

\item{active}{If `TRUE`, use only spots selected by `ActiveIdent()`}

\item{direction}{If `1` or `-1`, subset region 1 or 0 respectively. Default is `NULL`.}

\item{regress_out_cc}{If `TRUE`, regress out the curve parameter `t`}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A `SENNA` object with peak SVGs stored in `Gene$PeakSVGs`
}
\description{
Identify genes with peak-like spatial patterns along the C–S distance  
in regionation analysis, optionally controlling for curve parameter.
}
