% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGs_tools.R
\name{midpregi}
\alias{midpregi}
\title{Generics for Regionation Peak Detection Model}
\usage{
midpregi(senna, ...)
}
\arguments{
\item{senna}{A `SENNA` or `mSENNA` object}

\item{...}{Additional arguments}
}
\description{
Define generic function `midpregi()`, which identifies genes with peak-like  
expression patterns along the regionation axis.
}
