% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_senna.R
\docType{class}
\name{mSENNA-class}
\alias{mSENNA-class}
\title{mSENNA: S4 class for multi-SENNA approach}
\description{
Defines the S4 class 'mSENNA' for handling multiple SENNA objects.
}
\section{Slots}{

\describe{
\item{\code{SENNA}}{A list of SENNA objects}

\item{\code{msR}}{A report object for multi-SENNA analysis}
}}

