% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_senna.R
\docType{class}
\name{msR-class}
\alias{msR-class}
\title{msR: S4 class for the output report of multi-SENNA analysis}
\description{
Defines the S4 'msR' class, an output object for multi-SENNA analysis.
}
\section{Slots}{

\describe{
\item{\code{Report}}{A data frame containing the analysis report}

\item{\code{Variable_gene}}{A list of two character vectors storing significant genes with increasing and decreasing expression, respectively}

\item{\code{Filtered_gene}}{A character vector of genes filtered out due to singularity}
}}

