% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{PeakVolPlot}
\alias{PeakVolPlot}
\title{Volcano plot for peak-based SVGs}
\usage{
PeakVolPlot(senna, ...)
}
\arguments{
\item{senna}{A `SENNA` or `mSENNA` object}

\item{...}{Additional arguments passed to methods}
}
\value{
A volcano plot or list of plots highlighting peak-based SVGs
}
\description{
Define generic function `PeakVolPlot()`.
}
\seealso{
\code{\link{PeakVolPlot,SENNA-method}}
}
