% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{plug_coef}
\alias{plug_coef}
\title{Evaluate spline curve at a given parameter}
\usage{
plug_coef(t, coef_mat)
}
\arguments{
\item{t}{A curve parameter (numeric)}

\item{coef_mat}{A coefficient matrix for a cubic spline (each row: 4 coefficients for a segment)}
}
\value{
A numeric value representing the spatial coordinate at curve parameter `t`
}
\description{
Restores the spatial coordinate by evaluating a cubic spline segment at a given curve parameter `t`, using the provided coefficient matrix.
}
