% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGs_tools.R
\name{ProgSVGs,mSENNA-method}
\alias{ProgSVGs,mSENNA-method}
\title{SVGs for Progression Analysis}
\usage{
\S4method{ProgSVGs}{mSENNA}(
  senna,
  intervals,
  weight = "box",
  FDR_level = 0.01,
  grad_cutoff = 0.1,
  intensity = 1,
  active = FALSE,
  cores = 1L
)
}
\arguments{
\item{senna}{A mSENNA object.}

\item{intervals}{A numeric vector of intervals for progression analysis}

\item{weight}{`"gaussian"` or `"box"` weighting scheme}

\item{FDR_level}{FDR threshold (default = 0.01)}

\item{grad_cutoff}{Threshold for the absolute value of regression coefficients}

\item{intensity}{Intensity of Gaussian weights}

\item{active}{If `TRUE`, only include cells selected by `ActiveIdent()`}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A `msR` (multi-SENNA report) object
}
\description{
Identify progression-based spatially variable genes (SVGs) across multiple SENNA objects,
using weighted regression on the C–S distance. Supports Gaussian or box-shaped weighting schemes
and batch correction via factor modeling.
}
\seealso{
\code{\link[=ProgSVGs]{ProgSVGs generic}}, 
         \code{\link[=ProgSVGs,SENNA-method]{ProgSVGs for SENNA}}
}
