% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGs_tools.R
\name{ProgSVGs,SENNA-method}
\alias{ProgSVGs,SENNA-method}
\title{SVGs for Progression Analysis}
\usage{
\S4method{ProgSVGs}{SENNA}(
  senna,
  weight = "box",
  FDR_level = 0.01,
  grad_cutoff = 0.1,
  active = FALSE,
  interval = sqrt(2),
  intensity = 1,
  cores = 1L
)
}
\arguments{
\item{senna}{A SENNA object}

\item{weight}{`"gaussian"` or `"box"` weighting scheme}

\item{FDR_level}{FDR threshold}

\item{grad_cutoff}{Threshold for the absolute value of regression coefficients}

\item{active}{If `TRUE`, only include cells selected by `ActiveIdent()`}

\item{interval}{Size of interval for sliding window (used in box or Gaussian weighting)}

\item{intensity}{Intensity of Gaussian weights}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A SENNA object with progression SVGs stored in `Gene$P.SVGs`
}
\description{
Identify progression-based spatially variable genes (SVGs) using weighted regression on the C–S distance.
Supports Gaussian or box-shaped weighting and optional spot selection via `ActiveIdent()`.
}
\seealso{
\code{\link[=ProgSVGs]{ProgSVGs generic}}, 
         \code{\link[=ProgSVGs,mSENNA-method]{ProgSVGs for mSENNA}}
}
