% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGs_tools.R
\name{ProgSVGs}
\alias{ProgSVGs}
\title{Generics for Progression Analysis}
\usage{
ProgSVGs(senna, ...)
}
\arguments{
\item{senna}{A SENNA or mSENNA object}

\item{...}{Additional arguments}
}
\value{
A plot with progression SVGs.
}
\description{
Define generic function `ProgSVGs()`.
}
\seealso{
\code{\link[=ProgSVGs,SENNA-method]{ProgSVGs for SENNA}}, 
         \code{\link[=ProgSVGs,mSENNA-method]{ProgSVGs for mSENNA}}
}
