% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ProgVolPlot}
\alias{ProgVolPlot}
\title{Volcano plot for progression SVGs}
\usage{
ProgVolPlot(senna, ...)
}
\arguments{
\item{senna}{A `SENNA` or `mSENNA` object}

\item{...}{Additional arguments passed to the corresponding method}
}
\value{
A `ggplot` object representing a volcano plot of progression SVGs
}
\description{
Define the generic function `ProgVolPlot()` to visualize progression SVGs  
from either a `SENNA` or `mSENNA` object. See method implementations for details.
}
\seealso{
\code{\link[=ProgVolPlot,SENNA-method]{ProgVolPlot for SENNA}},  
  \code{\link[=ProgVolPlot,msR-method]{ProgVolPlot for msR}}
}
