% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGs_tools.R
\name{psvgs_box}
\alias{psvgs_box}
\title{Internal generic for weighted progression analysis (Box)}
\usage{
psvgs_box(senna, ...)
}
\arguments{
\item{senna}{A SENNA object}

\item{...}{Additional arguments}
}
\description{
Define generic function `psvgs_box()`. Used internally by `ProgSVGs()`.
}
