% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGs_tools.R
\name{psvgs_gwei,mSENNA-method}
\alias{psvgs_gwei,mSENNA-method}
\title{Multi-progression SVGs with Gaussian weights}
\usage{
\S4method{psvgs_gwei}{mSENNA}(senna, FDR_level, grad_cutoff, active, intervals, intensity, cores)
}
\arguments{
\item{senna}{A \code{mSENNA} object containing multiple SENNA instances.}

\item{FDR_level}{False discovery rate threshold.}

\item{grad_cutoff}{Threshold for absolute value of regression coefficients (gradients).}

\item{active}{If \code{TRUE}, only use activated spots defined by \code{ActiveIdent()}.}

\item{intervals}{A numeric vector specifying the maximum distance for each SENNA object to include in the regression.}

\item{intensity}{A numeric value scaling the Gaussian kernel.}

\item{cores}{Number of CPU cores to use (not supported on Windows).}
}
\value{
An \code{msR} object containing:  
\itemize{
  \item \code{Report}: Full regression statistics for each gene
  \item \code{Variable_gene}: A list of positively and negatively correlated genes
  \item \code{Filtered_gene}: Genes excluded from testing due to low expression
}
}
\description{
Identify spatially variable genes (SVGs) from multiple SENNA objects using Gaussian-weighted regression on the curve axis.  
The method assigns a separate regression model for each gene, weighted by a Gaussian kernel centered at each spot.  
Distance cutoffs and kernel intensities are defined per SENNA object.
}
