% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGs_tools.R
\name{psvgs_gwei}
\alias{psvgs_gwei}
\title{Internal generic for weighted progression analysis (Gaussian)}
\usage{
psvgs_gwei(senna, ...)
}
\arguments{
\item{senna}{A SENNA object}

\item{...}{Additional arguments}
}
\description{
Define generic function `psvgs_gwei()`. Used internally by `ProgSVGs()`.
}
