% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGs_tools.R
\name{RegionSVGs,mSENNA-method}
\alias{RegionSVGs,mSENNA-method}
\title{SVGs for Regionation Analysis}
\usage{
\S4method{RegionSVGs}{mSENNA}(
  senna,
  FDR_level = 0.01,
  grad_cutoff = 0.1,
  active = FALSE,
  direction = NULL,
  regress_out_cc = FALSE,
  cores = 1L
)
}
\arguments{
\item{senna}{A `mSENNA` object}

\item{FDR_level}{FDR threshold (default = 0.01)}

\item{grad_cutoff}{Threshold for the absolute value of regression coefficients}

\item{active}{If `TRUE`, only use spots selected by `ActiveIdent()`}

\item{direction}{A numeric vector specifying which region to include per SENNA object (e.g., c(1, -1, NULL))}

\item{regress_out_cc}{If `TRUE`, regress out the curve parameter `t`}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A `mSENNA` object with updated `msR` slot containing regionation SVG results
}
\description{
Identify region-based spatially variable genes (SVGs) from multiple SENNA objects.  
Performs regression on C–S distance (optionally adjusting for curve parameter and batch) and aggregates results into an `msR` object.
}
\seealso{
\code{\link[=RegionSVGs]{RegionSVGs generic}}, 
         \code{\link[=RegionSVGs,SENNA-method]{RegionSVGs for SENNA}}
}
