% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGs_tools.R
\name{RegionSVGs,SENNA-method}
\alias{RegionSVGs,SENNA-method}
\title{SVGs for Regionation Analysis}
\usage{
\S4method{RegionSVGs}{SENNA}(
  senna,
  FDR_level = 0.01,
  grad_cutoff = 0.1,
  active = FALSE,
  direction = NULL,
  regress_out_cc = FALSE,
  cores = 1L
)
}
\arguments{
\item{senna}{A SENNA object}

\item{FDR_level}{FDR threshold (default = 0.01)}

\item{grad_cutoff}{Threshold for the absolute value of regression coefficients}

\item{active}{If `TRUE`, only use spots selected by `ActiveIdent()`}

\item{direction}{If `1` or `-1`, only include spots in that region. Default is `NULL`.}

\item{regress_out_cc}{If `TRUE`, regress out the curve parameter `t`}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A SENNA object with regionation SVGs stored in `Gene$R.SVGs`
}
\description{
Identify region-based spatially variable genes (SVGs) using linear regression on the C–S distance.  
Optionally includes spots filtered by `ActiveIdent()` and can regress out the curve parameter.
}
\seealso{
\code{\link[=RegionSVGs]{RegionSVGs generic}}, 
         \code{\link[=RegionSVGs,mSENNA-method]{RegionSVGs for mSENNA}}
}
