% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGs_tools.R
\name{RegionSVGs}
\alias{RegionSVGs}
\title{Generic for regionation-based SVG analysis}
\usage{
RegionSVGs(senna, ...)
}
\arguments{
\item{senna}{A SENNA or mSENNA object}

\item{...}{Additional arguments}
}
\value{
A plot showing region-based spatially variable genes (SVGs).
}
\description{
Defines a generic for identifying or visualizing region-based spatially variable genes (SVGs).
}
\seealso{
\code{\link[=RegionSVGs,SENNA-method]{RegionSVGs for SENNA}}, 
         \code{\link[=RegionSVGs,mSENNA-method]{RegionSVGs for mSENNA}}
}
