% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{RegionVolPlot}
\alias{RegionVolPlot}
\title{Volcano plot for regionation SVGs}
\usage{
RegionVolPlot(senna, ...)
}
\arguments{
\item{senna}{A `SENNA` or `mSENNA` object}

\item{...}{Additional arguments passed to the method}
}
\value{
A `ggplot` object representing the volcano plot
}
\description{
Define generic function `RegionVolPlot`.
}
\seealso{
\code{\link[=RegionVolPlot,SENNA-method]{RegionVolPlot for SENNA}},
  \code{\link[=RegionVolPlot,msR-method]{RegionVolPlot for msR}}
}
