% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{scan_a_cell_lin}
\alias{scan_a_cell_lin}
\title{Find curve parameter for a single point (full linear curve)}
\usage{
scan_a_cell_lin(point, xcoef, ycoef, cores = 1L)
}
\arguments{
\item{point}{A 2D spatial coordinate (x, y)}

\item{xcoef}{Polynomial coefficients for x(t)}

\item{ycoef}{Polynomial coefficients for y(t)}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A numeric vector of length 2: estimated curve parameter (t) and squared distance
}
\description{
Estimates the curve parameter (t) for a spatial point by minimizing squared distance to a piecewise linear curve axis (CA), using polynomial root solving.
}
