% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{scan_a_celli}
\alias{scan_a_celli}
\title{Find curve parameter for a single point (islet curve)}
\usage{
scan_a_celli(point, hfun, xcoef, xder, ycoef, yder, init, cores = 1L)
}
\arguments{
\item{point}{A 2D spatial coordinate (x, y)}

\item{hfun}{A list of h(t) polynomial terms for each curve segment}

\item{xcoef}{Polynomial coefficients for x(t)}

\item{xder}{Derivatives of x(t)}

\item{ycoef}{Polynomial coefficients for y(t)}

\item{yder}{Derivatives of y(t)}

\item{init}{The initial knot coordinate used to close}

\item{cores}{Number of cores for parallel computation (not supported on Windows)}
}
\value{
A numeric vector of length 2: estimated curve parameter (t) and squared distance
}
\description{
Estimates the curve parameter (t) for a spatial point by solving h(t) = ⟨r - γ(t), γ'(t)⟩ = 0 along an islet-type closed curve axis (CA).  
Also considers the distance from the initial knot to enforce closed-curve continuity.
}
