% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_senna.R
\docType{class}
\name{SENNA-class}
\alias{SENNA-class}
\title{SENNA: S4 class for user-interactive spatial analysis}
\description{
Defines the S4 class 'SENNA' for user-interactive spatial analysis.
}
\section{Slots}{

\describe{
\item{\code{Gene}}{Gene count data}

\item{\code{Coord}}{Spatial coordinate data}

\item{\code{CurveAxis}}{Curve axis function}
}}

