% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_senna.R
\name{SENNA_Visium}
\alias{SENNA_Visium}
\title{Create SENNA object from Visium dataset}
\usage{
SENNA_Visium(
  InputObj,
  assay = NULL,
  slice_name,
  all_genes = FALSE,
  annotation = FALSE
)
}
\arguments{
\item{InputObj}{Input Seurat object. Must be normalized (e.g., via SCTransform).}

\item{assay}{Assay name}

\item{slice_name}{Slice name of the input Seurat object (must match exactly)}

\item{all_genes}{Logical. If TRUE, uses all features instead of only variable ones}

\item{annotation}{Logical. If TRUE, includes cell annotation in the output object}
}
\value{
A `SENNA` object
}
\description{
Converts a 10X Visium `Seurat` object into a `SENNA` object.
}
