% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_senna.R
\name{SENNA_Xenium}
\alias{SENNA_Xenium}
\title{Create SENNA object from Xenium dataset}
\usage{
SENNA_Xenium(InputObj, assay = NULL, fov_name, annotation = FALSE)
}
\arguments{
\item{InputObj}{Input Seurat object. Must be normalized (e.g., via SCTransform).}

\item{assay}{Assay name}

\item{fov_name}{FOV name of the input Seurat object (must match exactly)}

\item{annotation}{Logical. If TRUE, includes cell annotation in the output object}
}
\value{
A `SENNA` object
}
\description{
Converts a 10X Xenium `Seurat` object into a `SENNA` object.
}
