% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ShowBins}
\alias{ShowBins}
\title{Show spatial bins along the curve axis}
\usage{
ShowBins(senna, ...)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{...}{Additional arguments passed to the method}
}
\description{
Define generic function `ShowBins()` to visualize spatial bins  
segmented along the fitted curve axis.
}
\seealso{
\code{\link[=ShowBins,SENNA-method]{ShowBins for SENNA}}
}
