% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ShowCSDistance,SENNA-method}
\alias{ShowCSDistance,SENNA-method}
\title{Plot C-S distance along the curve axis}
\usage{
\S4method{ShowCSDistance}{SENNA}(
  senna,
  dot_size = 1,
  dot_alpha = 1,
  line_size = 1,
  high = "#98499a",
  medium = "#bbbbbb",
  low = "#67b665",
  bgcol = "#dddddd",
  direction = NULL,
  set_midpoint = 0
)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{dot_size}{Size of spot dots}

\item{dot_alpha}{Transparency of spot dots}

\item{line_size}{Size of the curve axis trace}

\item{high}{Color for high (positive) distance}

\item{medium}{Color for zero or near-zero distance}

\item{low}{Color for low (negative) distance}

\item{bgcol}{Color for background spots (if `direction` is specified)}

\item{direction}{Optional region ID to highlight (`1` or `-1`).
Spots not in this region are shown in background color.}

\item{set_midpoint}{The midpoint value used for `scale_colour_gradient2()` (only when `direction = NULL`)}
}
\value{
A `ggplot` object showing spatial distribution of C-S distances
}
\description{
Visualize each spot’s signed or absolute distance from the curve axis.  
Optionally highlights one side of the region by showing the other in background color.
}
