% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ShowCSDistance}
\alias{ShowCSDistance}
\title{generics for CSD plot}
\usage{
ShowCSDistance(senna, ...)
}
\arguments{
\item{senna}{A `SENNA` or `mSENNA` object}

\item{...}{Additional arguments passed to the method}
}
\value{
A `ggplot` object visualizing C-S distance
}
\description{
Define generic function `ShowCSDistance()` for plotting C-S distance from the curve axis.  
Methods are implemented for both `SENNA` and `mSENNA` objects.
}
\seealso{
\code{\link[=ShowCSDistance,SENNA-method]{ShowCSDistance for SENNA}},  
  \code{\link[=ShowCSDistance,mSENNA-method]{ShowCSDistance for mSENNA}}
}
