% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ShowCurve}
\alias{ShowCurve}
\title{Visualize the curve axis from a SENNA object.}
\usage{
ShowCurve(senna, ...)
}
\arguments{
\item{senna}{A SENNA object}

\item{...}{Additional arguments passed to the method}
}
\description{
Defines the generic `ShowCurve()` function for plotting the fitted curve axis  
along with background spatial coordinates. See method documentation for specific plotting behavior.
}
\seealso{
\code{\link[=ShowCurve,SENNA-method]{ShowCurve for SENNA}}
}
