% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_senna.R
\docType{class}
\name{simplesktS-class}
\alias{simplesktS-class}
\title{simplesktS: S4 class for visualizing SENNA object}
\description{
Defines the S4 class 'simplesktS' (Sketch SENNA).
}
\section{Slots}{

\describe{
\item{\code{sketch}}{A matrix generated by `SketchPatterns()`}

\item{\code{label}}{A character vector of gene names}

\item{\code{n_spots}}{An integer vector indicating the number of spots for each value}
}}

