% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{SketchPatterns}
\alias{SketchPatterns}
\title{Sketch specific pattern for each SVG}
\usage{
SketchPatterns(senna, ...)
}
\arguments{
\item{senna}{A `SENNA`, `mSENNA`, or `msR` object}

\item{...}{Additional arguments passed to methods}
}
\description{
Define generic function `SketchPatterns`.
}
\seealso{
\code{\link[=SketchPatterns,SENNA-method]{SketchPatterns for SENNA}},
  \code{\link[=SketchPatterns,mSENNA-method]{SketchPatterns for mSENNA}}
}
