% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_senna.R
\docType{class}
\name{sktS-class}
\alias{sktS-class}
\title{sktS: S4 class for visualizing SENNA object}
\description{
Defines the S4 class 'sktS' (Sketch SENNA).
}
\section{Slots}{

\describe{
\item{\code{sketch}}{A data frame generated by `SketchPatterns()`}

\item{\code{label}}{A character vector of gene names}
}}

