% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{TissueFeaturePlot,SENNA-method}
\alias{TissueFeaturePlot,SENNA-method}
\title{Plot gene expression on tissue}
\usage{
\S4method{TissueFeaturePlot}{SENNA}(
  senna,
  gene,
  interval = NULL,
  active = FALSE,
  direction = NULL,
  curve_axes = FALSE,
  low = "#cfcfcf",
  high = "#800000",
  size = 1,
  line_size = 1,
  alpha = 1
)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{gene}{Name of a gene to visualize}

\item{interval}{Optional numeric distance threshold for inclusion based on C-S distance}

\item{active}{Whether to restrict to activated cells (default = FALSE)}

\item{direction}{Optional region direction filter (e.g., 1 or -1)}

\item{curve_axes}{If TRUE, overlays the curve axis (default = FALSE)}

\item{low}{Color for low expression values}

\item{high}{Color for high expression values}

\item{size}{Size of the gene expression points}

\item{line_size}{Size of the curve axis overlay}

\item{alpha}{Transparency level for points (0–1)}
}
\value{
A `ggplot` object showing gene expression over the tissue
}
\description{
Visualize the expression of a specific gene on the spatial coordinate system  
of a `SENNA` object. Supports filtering by activation status, region, or  
proximity to curve axis (C-S distance), and optional overlay of the curve axis.
}
\seealso{
\code{\link{TissueFeaturePlot}}
}
