% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{TissueFeaturePlot}
\alias{TissueFeaturePlot}
\title{Tissue feature plot generic}
\usage{
TissueFeaturePlot(senna, gene, ...)
}
\arguments{
\item{senna}{A `SENNA` or `mSENNA` object}

\item{gene}{Name of gene to plot}

\item{...}{Additional arguments passed to method implementations}
}
\value{
A `ggplot` object showing spatial gene expression
}
\description{
Define the generic function for `TissueFeaturePlot()`, which visualizes  
gene expression over spatial coordinates in `SENNA` or `mSENNA` objects.
}
\seealso{
\code{\link{TissueFeaturePlot,SENNA-method}}, \code{\link{TissueFeaturePlot,mSENNA-method}}
}
