% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{TissueFOV}
\alias{TissueFOV}
\title{Visualize spatial tissue layout from a SENNA object.}
\usage{
TissueFOV(senna, dot_size = 1, dot_alpha = 1, colors = NULL)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{dot_size}{Numeric value for dot size (default = 1)}

\item{dot_alpha}{Numeric value for dot transparency (default = 1)}

\item{colors}{Optional named vector of colors for manual cell type coloring}
}
\value{
A `ggplot` object representing the tissue spatial layout
}
\description{
Plot spatial coordinates from a `SENNA` object to visualize the tissue field-of-view (FOV).  
If annotation exists in `Gene$Reference$Annotation`, color spots by cell type.
}
