% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_axis.R
\name{TrimmedCurve,SENNA-method}
\alias{TrimmedCurve,SENNA-method}
\title{Generate trimmed curve axis}
\usage{
\S4method{TrimmedCurve}{SENNA}(senna, knot_df, type = "spline")
}
\arguments{
\item{senna}{A `SENNA` object}

\item{knot_df}{A data frame of knots}

\item{type}{Type of curve. Must be one of `'spline'`, `'straight'`, or `'islet'`}
}
\value{
A `SENNA` object with curve axis added
}
\description{
Method for the generic `TrimmedCurve()` applied to a `SENNA` object.
}
