% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_axis.R
\name{TrimmedCurve}
\alias{TrimmedCurve}
\title{Create a Curve Axis (CA) by Connecting the Given Knots without Extrapolation}
\usage{
TrimmedCurve(senna, ...)
}
\arguments{
\item{senna}{A SENNA object}

\item{...}{Additional arguments}
}
\description{
Defines the generic of `TrimmedCurve()`. It is suitable for both 'islet' scenarios and 'progression' scenarios where only part of the FOV is used.
}
