% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate.R
\name{trplug_coef_lin}
\alias{trplug_coef_lin}
\title{Evaluate trimmed linear curve at a given parameter}
\usage{
trplug_coef_lin(t, coef_mat)
}
\arguments{
\item{t}{A curve parameter (numeric)}

\item{coef_mat}{A coefficient matrix for linear segments (each row: slope and intercept)}
}
\value{
A numeric value representing the spatial coordinate at curve parameter `t`
}
\description{
Restores the spatial coordinate by evaluating a trimmed piecewise linear curve axis (CA) at a given curve parameter `t`.
}
