% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_axis.R
\name{trtype_islet}
\alias{trtype_islet}
\title{Trimmed curve (islet)}
\usage{
trtype_islet(senna, knot_df)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{knot_df}{A data frame of knots}
}
\value{
A `SENNA` object with closed islet-shaped curve axis stored
}
\description{
Computes coefficients for a closed curve axis (CA) based on islet-shaped knots.
}
