% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_axis.R
\name{trtype_linear}
\alias{trtype_linear}
\title{Trimmed curve (piecewise linear)}
\usage{
trtype_linear(senna, knot_df, t)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{knot_df}{A data frame of knots}

\item{t}{A numeric vector specifying knot positions}
}
\value{
A `SENNA` object with trimmed linear curve axis stored
}
\description{
Computes piecewise linear trimmed curve axis (CA) coefficients from given knots.
}
