% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_axis.R
\name{trtype_spline}
\alias{trtype_spline}
\title{Trimmed curve (spline)}
\usage{
trtype_spline(senna, knot_df, t)
}
\arguments{
\item{senna}{A `SENNA` object}

\item{knot_df}{A data frame of knots}

\item{t}{A numeric vector specifying knot positions}
}
\value{
A `SENNA` object with trimmed spline curve axis stored
}
\description{
Computes spline-based trimmed curve axis (CA) coefficients from given knots.
}
